Fix ts-land
