+++ yarn-project/ethereum/src/deploy_l1_contracts.ts
@@ -3,53 +3,6 @@ import { EthAddress } from '@aztec/foundation/eth-address';
 import type { Fr } from '@aztec/foundation/fields';
 import { type Logger, createLogger } from '@aztec/foundation/log';
 import { DateProvider } from '@aztec/foundation/timer';
-import {
-  CoinIssuerAbi,
-  CoinIssuerBytecode,
-  ExtRollupLib2Abi,
-  ExtRollupLib2Bytecode,
-  ExtRollupLibAbi,
-  ExtRollupLibBytecode,
-  FeeAssetHandlerAbi,
-  FeeAssetHandlerBytecode,
-  FeeJuicePortalAbi,
-  FeeJuicePortalBytecode,
-  GSEAbi,
-  GSEBytecode,
-  GovernanceAbi,
-  GovernanceBytecode,
-  GovernanceProposerAbi,
-  GovernanceProposerBytecode,
-  HonkVerifierAbi,
-  HonkVerifierBytecode,
-  InboxAbi,
-  InboxBytecode,
-  MockVerifierAbi,
-  MockVerifierBytecode,
-  MockZKPassportVerifierAbi,
-  MockZKPassportVerifierBytecode,
-  MultiAdderAbi,
-  MultiAdderBytecode,
-  OutboxAbi,
-  OutboxBytecode,
-  RegisterNewRollupVersionPayloadAbi,
-  RegisterNewRollupVersionPayloadBytecode,
-  RegistryAbi,
-  RegistryBytecode,
-  RewardDistributorAbi,
-  RewardDistributorBytecode,
-  RollupAbi,
-  RollupBytecode,
-  RollupLinkReferences,
-  SlashFactoryAbi,
-  SlashFactoryBytecode,
-  StakingAssetHandlerAbi,
-  StakingAssetHandlerBytecode,
-  TestERC20Abi,
-  TestERC20Bytecode,
-  ValidatorSelectionLibAbi,
-  ValidatorSelectionLibBytecode,
-} from '@aztec/l1-artifacts';
 
 import type { Abi, Narrow } from 'abitype';
 import {
@@ -81,6 +34,23 @@ import {
 import { deployMulticall3 } from './contracts/multicall.js';
 import { RegistryContract } from './contracts/registry.js';
 import { RollupContract } from './contracts/rollup.js';
+import {
+  CoinIssuerArtifact,
+  FeeAssetArtifact,
+  FeeAssetHandlerArtifact,
+  GSEArtifact,
+  GovernanceArtifact,
+  GovernanceProposerArtifact,
+  MultiAdderArtifact,
+  RegisterNewRollupVersionPayloadArtifact,
+  RegistryArtifact,
+  RollupArtifact,
+  SlashFactoryArtifact,
+  StakingAssetArtifact,
+  StakingAssetHandlerArtifact,
+  l1ArtifactsVerifiers,
+  mockVerifiers,
+} from './l1_artifacts.js';
 import type { L1ContractAddresses } from './l1_contract_addresses.js';
 import {
   type GasPrice,
@@ -146,115 +116,6 @@ export interface ContractArtifacts {
   libraries?: Libraries;
 }
 
-export const l1Artifacts = {
-  registry: {
-    contractAbi: RegistryAbi,
-    contractBytecode: RegistryBytecode as Hex,
-  },
-  inbox: {
-    contractAbi: InboxAbi,
-    contractBytecode: InboxBytecode as Hex,
-  },
-  outbox: {
-    contractAbi: OutboxAbi,
-    contractBytecode: OutboxBytecode as Hex,
-  },
-  rollup: {
-    contractAbi: RollupAbi,
-    contractBytecode: RollupBytecode as Hex,
-    libraries: {
-      linkReferences: RollupLinkReferences,
-      libraryCode: {
-        ValidatorSelectionLib: {
-          contractAbi: ValidatorSelectionLibAbi,
-          contractBytecode: ValidatorSelectionLibBytecode as Hex,
-        },
-        ExtRollupLib: {
-          contractAbi: ExtRollupLibAbi,
-          contractBytecode: ExtRollupLibBytecode as Hex,
-        },
-        ExtRollupLib2: {
-          contractAbi: ExtRollupLib2Abi,
-          contractBytecode: ExtRollupLib2Bytecode as Hex,
-        },
-      },
-    },
-  },
-  stakingAsset: {
-    contractAbi: TestERC20Abi,
-    contractBytecode: TestERC20Bytecode as Hex,
-  },
-  feeAsset: {
-    contractAbi: TestERC20Abi,
-    contractBytecode: TestERC20Bytecode as Hex,
-  },
-  feeJuicePortal: {
-    contractAbi: FeeJuicePortalAbi,
-    contractBytecode: FeeJuicePortalBytecode as Hex,
-  },
-  rewardDistributor: {
-    contractAbi: RewardDistributorAbi,
-    contractBytecode: RewardDistributorBytecode as Hex,
-  },
-  coinIssuer: {
-    contractAbi: CoinIssuerAbi,
-    contractBytecode: CoinIssuerBytecode as Hex,
-  },
-  governanceProposer: {
-    contractAbi: GovernanceProposerAbi,
-    contractBytecode: GovernanceProposerBytecode as Hex,
-  },
-  governance: {
-    contractAbi: GovernanceAbi,
-    contractBytecode: GovernanceBytecode as Hex,
-  },
-  slashFactory: {
-    contractAbi: SlashFactoryAbi,
-    contractBytecode: SlashFactoryBytecode as Hex,
-  },
-  registerNewRollupVersionPayload: {
-    contractAbi: RegisterNewRollupVersionPayloadAbi,
-    contractBytecode: RegisterNewRollupVersionPayloadBytecode as Hex,
-  },
-  feeAssetHandler: {
-    contractAbi: FeeAssetHandlerAbi,
-    contractBytecode: FeeAssetHandlerBytecode as Hex,
-  },
-  stakingAssetHandler: {
-    contractAbi: StakingAssetHandlerAbi,
-    contractBytecode: StakingAssetHandlerBytecode as Hex,
-  },
-  multiAdder: {
-    contractAbi: MultiAdderAbi,
-    contractBytecode: MultiAdderBytecode as Hex,
-  },
-};
-
-// Moving this out to avoid "type too big" error.
-// Palla has a proper fix for this.
-export const gseArtifact = {
-  contractAbi: GSEAbi,
-  contractBytecode: GSEBytecode as Hex,
-};
-
-export const l1ArtifactsVerifiers = {
-  honkVerifier: {
-    contractAbi: HonkVerifierAbi,
-    contractBytecode: HonkVerifierBytecode as Hex,
-  },
-};
-
-const mockVerifiers = {
-  mockVerifier: {
-    contractAbi: MockVerifierAbi,
-    contractBytecode: MockVerifierBytecode as Hex,
-  },
-  mockZkPassportVerifier: {
-    contractAbi: MockZKPassportVerifierAbi,
-    contractBytecode: MockZKPassportVerifierBytecode as Hex,
-  },
-};
-
 export interface DeployL1ContractsArgs extends L1ContractsConfig {
   /** The vk tree root. */
   vkTreeRoot: Fr;
@@ -297,14 +158,14 @@ export const deploySharedContracts = async (
 
   const txHashes: Hex[] = [];
 
-  const feeAssetAddress = await deployer.deploy(l1Artifacts.feeAsset, [
+  const feeAssetAddress = await deployer.deploy(FeeAssetArtifact, [
     'FeeJuice',
     'FEE',
     l1Client.account.address.toString(),
   ]);
   logger.verbose(`Deployed Fee Asset at ${feeAssetAddress}`);
 
-  const stakingAssetAddress = await deployer.deploy(l1Artifacts.stakingAsset, [
+  const stakingAssetAddress = await deployer.deploy(StakingAssetArtifact, [
     'Staking',
     'STK',
     l1Client.account.address.toString(),
@@ -313,7 +174,7 @@ export const deploySharedContracts = async (
 
   const gseConfiguration = getGSEConfiguration(networkName);
 
-  const gseAddress = await deployer.deploy(gseArtifact, [
+  const gseAddress = await deployer.deploy(GSEArtifact, [
     l1Client.account.address.toString(),
     stakingAssetAddress.toString(),
     gseConfiguration.depositAmount,
@@ -321,13 +182,13 @@ export const deploySharedContracts = async (
   ]);
   logger.verbose(`Deployed GSE at ${gseAddress}`);
 
-  const registryAddress = await deployer.deploy(l1Artifacts.registry, [
+  const registryAddress = await deployer.deploy(RegistryArtifact, [
     l1Client.account.address.toString(),
     feeAssetAddress.toString(),
   ]);
   logger.verbose(`Deployed Registry at ${registryAddress}`);
 
-  const governanceProposerAddress = await deployer.deploy(l1Artifacts.governanceProposer, [
+  const governanceProposerAddress = await deployer.deploy(GovernanceProposerArtifact, [
     registryAddress.toString(),
     gseAddress.toString(),
     args.governanceProposerQuorum,
@@ -337,7 +198,7 @@ export const deploySharedContracts = async (
 
   // @note @LHerskind the assets are expected to be the same at some point, but for better
   // configurability they are different for now.
-  const governanceAddress = await deployer.deploy(l1Artifacts.governance, [
+  const governanceAddress = await deployer.deploy(GovernanceArtifact, [
     stakingAssetAddress.toString(),
     governanceProposerAddress.toString(),
     gseAddress.toString(),
@@ -353,7 +214,7 @@ export const deploySharedContracts = async (
   } else {
     const gseContract = getContract({
       address: getAddress(gseAddress.toString()),
-      abi: gseArtifact.contractAbi,
+      abi: GSEArtifact.contractAbi,
       client: l1Client,
     });
     const existingGovernance = await gseContract.read.getGovernance();
@@ -367,7 +228,7 @@ export const deploySharedContracts = async (
       {
         to: gseAddress.toString(),
         data: encodeFunctionData({
-          abi: gseArtifact.contractAbi,
+          abi: GSEArtifact.contractAbi,
           functionName: 'setGovernance',
           args: [governanceAddress.toString()],
         }),
@@ -379,7 +240,7 @@ export const deploySharedContracts = async (
     txHashes.push(txHash);
   }
 
-  const coinIssuerAddress = await deployer.deploy(l1Artifacts.coinIssuer, [
+  const coinIssuerAddress = await deployer.deploy(CoinIssuerArtifact, [
     feeAssetAddress.toString(),
     1n * 10n ** 18n, // @todo  #8084
     governanceAddress.toString(),
@@ -388,7 +249,7 @@ export const deploySharedContracts = async (
 
   const feeAsset = getContract({
     address: feeAssetAddress.toString(),
-    abi: l1Artifacts.feeAsset.contractAbi,
+    abi: FeeAssetArtifact.contractAbi,
     client: l1Client,
   });
 
@@ -400,7 +261,7 @@ export const deploySharedContracts = async (
       {
         to: feeAssetAddress.toString(),
         data: encodeFunctionData({
-          abi: l1Artifacts.feeAsset.contractAbi,
+          abi: FeeAssetArtifact.contractAbi,
           functionName: 'addMinter',
           args: [coinIssuerAddress.toString()],
         }),
@@ -423,7 +284,7 @@ export const deploySharedContracts = async (
     /*                          CHEAT CODES START HERE                            */
     /* -------------------------------------------------------------------------- */
 
-    feeAssetHandlerAddress = await deployer.deploy(l1Artifacts.feeAssetHandler, [
+    feeAssetHandlerAddress = await deployer.deploy(FeeAssetHandlerArtifact, [
       l1Client.account.address,
       feeAssetAddress.toString(),
       BigInt(1e18),
@@ -433,7 +294,7 @@ export const deploySharedContracts = async (
     const { txHash } = await deployer.sendTransaction({
       to: feeAssetAddress.toString(),
       data: encodeFunctionData({
-        abi: l1Artifacts.feeAsset.contractAbi,
+        abi: FeeAssetArtifact.contractAbi,
         functionName: 'addMinter',
         args: [feeAssetHandlerAddress.toString()],
       }),
@@ -466,15 +327,13 @@ export const deploySharedContracts = async (
         skipMerkleCheck: true, // skip merkle check - needed for testing without generating proofs
       };
 
-      stakingAssetHandlerAddress = await deployer.deploy(l1Artifacts.stakingAssetHandler, [
-        stakingAssetHandlerDeployArgs,
-      ]);
+      stakingAssetHandlerAddress = await deployer.deploy(StakingAssetHandlerArtifact, [stakingAssetHandlerDeployArgs]);
       logger.verbose(`Deployed StakingAssetHandler at ${stakingAssetHandlerAddress}`);
 
       const { txHash: stakingMinterTxHash } = await deployer.sendTransaction({
         to: stakingAssetAddress.toString(),
         data: encodeFunctionData({
-          abi: l1Artifacts.stakingAsset.contractAbi,
+          abi: StakingAssetArtifact.contractAbi,
           functionName: 'addMinter',
           args: [stakingAssetHandlerAddress.toString()],
         }),
@@ -510,7 +369,7 @@ export const deploySharedContracts = async (
   const { txHash: fundRewardDistributorTxHash } = await deployer.sendTransaction({
     to: feeAssetAddress.toString(),
     data: encodeFunctionData({
-      abi: l1Artifacts.feeAsset.contractAbi,
+      abi: FeeAssetArtifact.contractAbi,
       functionName: 'mint',
       args: [rewardDistributorAddress.toString(), funding],
     }),
@@ -592,7 +451,7 @@ export const deployRollupForUpgrade = async (
 };
 
 export const deploySlashFactory = async (deployer: L1Deployer, rollupAddress: Hex, logger: Logger) => {
-  const slashFactoryAddress = await deployer.deploy(l1Artifacts.slashFactory, [rollupAddress]);
+  const slashFactoryAddress = await deployer.deploy(SlashFactoryArtifact, [rollupAddress]);
   logger.verbose(`Deployed SlashFactory at ${slashFactoryAddress}`);
   return slashFactoryAddress;
 };
@@ -601,7 +460,7 @@ export const deployUpgradePayload = async (
   deployer: L1Deployer,
   addresses: Pick<L1ContractAddresses, 'registryAddress' | 'rollupAddress'>,
 ) => {
-  const payloadAddress = await deployer.deploy(l1Artifacts.registerNewRollupVersionPayload, [
+  const payloadAddress = await deployer.deploy(RegisterNewRollupVersionPayloadArtifact, [
     addresses.registryAddress.toString(),
     addresses.rollupAddress.toString(),
   ]);
@@ -682,7 +541,7 @@ export const deployRollup = async (
     rollupConfigArgs,
   ];
 
-  const rollupAddress = await deployer.deploy(l1Artifacts.rollup, rollupArgs);
+  const rollupAddress = await deployer.deploy(RollupArtifact, rollupArgs);
   logger.verbose(`Deployed Rollup at ${rollupAddress}`, rollupConfigArgs);
 
   const rollupContract = new RollupContract(extendedClient, rollupAddress);
@@ -697,7 +556,7 @@ export const deployRollup = async (
     const { txHash: mintTxHash } = await deployer.sendTransaction({
       to: addresses.feeJuiceAddress.toString(),
       data: encodeFunctionData({
-        abi: l1Artifacts.feeAsset.contractAbi,
+        abi: FeeAssetArtifact.contractAbi,
         functionName: 'mint',
         args: [feeJuicePortalAddress.toString(), args.feeJuicePortalInitialBalance],
       }),
@@ -708,13 +567,13 @@ export const deployRollup = async (
     txHashes.push(mintTxHash);
   }
 
-  const slashFactoryAddress = await deployer.deploy(l1Artifacts.slashFactory, [rollupAddress.toString()]);
+  const slashFactoryAddress = await deployer.deploy(SlashFactoryArtifact, [rollupAddress.toString()]);
   logger.verbose(`Deployed SlashFactory at ${slashFactoryAddress}`);
 
   // We need to call a function on the registry to set the various contract addresses.
   const registryContract = getContract({
     address: getAddress(addresses.registryAddress.toString()),
-    abi: l1Artifacts.registry.contractAbi,
+    abi: RegistryArtifact.contractAbi,
     client: extendedClient,
   });
 
@@ -728,7 +587,7 @@ export const deployRollup = async (
       const { txHash: addRollupTxHash } = await deployer.sendTransaction({
         to: addresses.registryAddress.toString(),
         data: encodeFunctionData({
-          abi: l1Artifacts.registry.contractAbi,
+          abi: RegistryArtifact.contractAbi,
           functionName: 'addRollup',
           args: [getAddress(rollupContract.address)],
         }),
@@ -746,7 +605,7 @@ export const deployRollup = async (
   // We need to call a function on the registry to set the various contract addresses.
   const gseContract = getContract({
     address: getAddress(addresses.gseAddress.toString()),
-    abi: gseArtifact.contractAbi,
+    abi: GSEArtifact.contractAbi,
     client: extendedClient,
   });
   if ((await gseContract.read.owner()) === getAddress(extendedClient.account.address)) {
@@ -754,7 +613,7 @@ export const deployRollup = async (
       const { txHash: addRollupTxHash } = await deployer.sendTransaction({
         to: addresses.gseAddress.toString(),
         data: encodeFunctionData({
-          abi: gseArtifact.contractAbi,
+          abi: GSEArtifact.contractAbi,
           functionName: 'addRollup',
           args: [getAddress(rollupContract.address)],
         }),
@@ -801,13 +660,13 @@ export const handoverToGovernance = async (
   // We need to call a function on the registry to set the various contract addresses.
   const registryContract = getContract({
     address: getAddress(registryAddress.toString()),
-    abi: l1Artifacts.registry.contractAbi,
+    abi: RegistryArtifact.contractAbi,
     client: extendedClient,
   });
 
   const gseContract = getContract({
     address: getAddress(gseAddress.toString()),
-    abi: gseArtifact.contractAbi,
+    abi: GSEArtifact.contractAbi,
     client: extendedClient,
   });
 
@@ -822,7 +681,7 @@ export const handoverToGovernance = async (
     const { txHash: transferOwnershipTxHash } = await deployer.sendTransaction({
       to: registryAddress.toString(),
       data: encodeFunctionData({
-        abi: l1Artifacts.registry.contractAbi,
+        abi: RegistryArtifact.contractAbi,
         functionName: 'transferOwnership',
         args: [getAddress(governanceAddress.toString())],
       }),
@@ -839,7 +698,7 @@ export const handoverToGovernance = async (
     const { txHash: transferOwnershipTxHash } = await deployer.sendTransaction({
       to: gseContract.address,
       data: encodeFunctionData({
-        abi: gseArtifact.contractAbi,
+        abi: GSEArtifact.contractAbi,
         functionName: 'transferOwnership',
         args: [getAddress(governanceAddress.toString())],
       }),
@@ -899,10 +758,7 @@ export const addMultipleValidators = async (
     }
 
     if (validators.length > 0) {
-      const multiAdder = await deployer.deploy(l1Artifacts.multiAdder, [
-        rollupAddress,
-        deployer.client.account.address,
-      ]);
+      const multiAdder = await deployer.deploy(MultiAdderArtifact, [rollupAddress, deployer.client.account.address]);
 
       const validatorsTuples = validators.map(v => ({
         attester: getAddress(v.attester.toString()),
@@ -914,7 +770,7 @@ export const addMultipleValidators = async (
       const { txHash } = await deployer.sendTransaction({
         to: stakingAssetAddress,
         data: encodeFunctionData({
-          abi: l1Artifacts.stakingAsset.contractAbi,
+          abi: StakingAssetArtifact.contractAbi,
           functionName: 'mint',
           args: [multiAdder.toString(), stakeNeeded],
         }),
@@ -927,7 +783,7 @@ export const addMultipleValidators = async (
 
       const addValidatorsTxHash = await deployer.client.writeContract({
         address: multiAdder.toString(),
-        abi: l1Artifacts.multiAdder.contractAbi,
+        abi: MultiAdderArtifact.contractAbi,
         functionName: 'addValidators',
         args: [validatorsTuples],
       });
@@ -959,7 +815,7 @@ export const cheat_initializeFeeAssetHandler = async (
   feeAssetHandlerAddress: EthAddress;
   txHash: Hex;
 }> => {
-  const feeAssetHandlerAddress = await deployer.deploy(l1Artifacts.feeAssetHandler, [
+  const feeAssetHandlerAddress = await deployer.deploy(FeeAssetHandlerArtifact, [
     extendedClient.account.address,
     feeAssetAddress.toString(),
     BigInt(1e18),
@@ -969,7 +825,7 @@ export const cheat_initializeFeeAssetHandler = async (
   const { txHash } = await deployer.sendTransaction({
     to: feeAssetAddress.toString(),
     data: encodeFunctionData({
-      abi: l1Artifacts.feeAsset.contractAbi,
+      abi: FeeAssetArtifact.contractAbi,
       functionName: 'addMinter',
       args: [feeAssetHandlerAddress.toString()],
     }),
