+++ yarn-project/ethereum/src/index.ts
@@ -12,3 +12,4 @@ export * from './queries.js';
 export * from './client.js';
 export * from './account.js';
 export * from './l1_types.js';
+export * from './l1_artifacts.js';
