+++ yarn-project/prover-node/src/prover-node-publisher.test.ts
@@ -138,6 +138,8 @@ describe('prover-node-publisher', () => {
       rollup.getBlock.mockImplementation((blockNumber: bigint) =>
         Promise.resolve({
           archive: blocks[Number(blockNumber) - 1].endArchiveRoot.toString(),
+          attestationsHash: '0x', // unused,
+          payloadDigest: '0x', // unused,
           headerHash: '0x', // unused,
           blobCommitmentsHash: '0x', // unused,
           slotNumber: 0n, // unused,
@@ -177,6 +179,7 @@ describe('prover-node-publisher', () => {
           publicInputs: ourPublicInputs,
           proof: Proof.empty(),
           batchedBlobInputs: ourBatchedBlob,
+          attestations: [],
         })
         .then(() => 'Success')
         .catch(error => error.message);
