+++ yarn-project/prover-node/src/prover-node-publisher.ts
@@ -1,6 +1,11 @@
 import { type BatchedBlob, FinalBlobAccumulatorPublicInputs } from '@aztec/blob-lib';
 import { AZTEC_MAX_EPOCH_DURATION } from '@aztec/constants';
-import type { L1TxUtils, RollupContract } from '@aztec/ethereum';
+import {
+  type L1TxUtils,
+  type RollupContract,
+  RollupContract as RollupContractClass,
+  type ViemCommitteeAttestation,
+} from '@aztec/ethereum';
 import { makeTuple } from '@aztec/foundation/array';
 import { areArraysEqual } from '@aztec/foundation/collection';
 import { EthAddress } from '@aztec/foundation/eth-address';
@@ -97,6 +102,7 @@ export class ProverNodePublisher {
     publicInputs: RootRollupPublicInputs;
     proof: Proof;
     batchedBlobInputs: BatchedBlob;
+    attestations: ViemCommitteeAttestation[];
   }): Promise<boolean> {
     const { epochNumber, fromBlock, toBlock } = args;
     const ctx = { epochNumber, fromBlock, toBlock };
@@ -150,6 +156,7 @@ export class ProverNodePublisher {
     publicInputs: RootRollupPublicInputs;
     proof: Proof;
     batchedBlobInputs: BatchedBlob;
+    attestations: ViemCommitteeAttestation[];
   }) {
     const { fromBlock, toBlock, publicInputs, batchedBlobInputs } = args;
 
@@ -207,6 +214,7 @@ export class ProverNodePublisher {
     publicInputs: RootRollupPublicInputs;
     proof: Proof;
     batchedBlobInputs: BatchedBlob;
+    attestations: ViemCommitteeAttestation[];
   }): Promise<TransactionReceipt | undefined> {
     const txArgs = [this.getSubmitEpochProofArgs(args)] as const;
 
@@ -246,6 +254,7 @@ export class ProverNodePublisher {
     toBlock: number;
     publicInputs: RootRollupPublicInputs;
     batchedBlobInputs: BatchedBlob;
+    attestations: ViemCommitteeAttestation[];
   }) {
     // Returns arguments for EpochProofLib.sol -> getEpochProofPublicInputs()
     return [
@@ -271,6 +280,7 @@ export class ProverNodePublisher {
     publicInputs: RootRollupPublicInputs;
     proof: Proof;
     batchedBlobInputs: BatchedBlob;
+    attestations: ViemCommitteeAttestation[];
   }) {
     // Returns arguments for EpochProofLib.sol -> submitEpochRootProof()
     const proofHex: Hex = `0x${args.proof.withoutPublicInputs().toString('hex')}`;
@@ -280,6 +290,7 @@ export class ProverNodePublisher {
       end: argsArray[1],
       args: argsArray[2],
       fees: argsArray[3],
+      attestations: RollupContractClass.packAttestations(args.attestations),
       blobInputs: argsArray[4],
       proof: proofHex,
     };
